sixth_version/backend/requirements.txt
aiofiles==24.1.0
asyncio==3.4.3
chroma-hnswlib==0.7.6
chromadb==1.0.12
fastapi==0.115.9
httpx==0.28.1
langchain==0.3.7
langchain-anthropic==0.3.15
langchain-chroma==0.2.4
langchain-community==0.3.7
lark==1.1.9
loguru==0.7.2
openai==1.78.0
pydantic==2.9.2
Pillow==10.3.0
pypdf==4.3.1
pymupdf4llm==0.0.25
PyMuPDF==1.26.1
python-dotenv==1.1.0
python-multipart==0.0.18
tiktoken==0.8.0
typing-extensions==4.12.2
uvicorn[standard]==0.30.6